requirements.txt
# Core dependencies
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
tqdm>=4.66.0
PyYAML>=6.0

# Streamlit app
streamlit>=1.28.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0